+++ lib/base/textedit.cpp
@@ -137,7 +137,7 @@ int TextEdit::readcurpos()
 QString TextEdit::readselected()
 {
   QString s=textCursor().selectedText();
-  s.replace(0x2029,'\n');
+  s.replace((short)0x2029,'\n');
   return s;
 }
